[ovsp4rt] Rename unit tests for consistency
