+++ ovs-p4rt/sidecar/testing/CMakeLists.txt
@@ -121,23 +121,23 @@ if(ES2K_TARGET)
 # ES2K unit tests
 #-----------------------------------------------------------------------
 
-define_ovsp4rt_test(fdb_rx_vlan_entry_test)
-define_ovsp4rt_test(fdb_smac_entry_test)
-define_ovsp4rt_test(fdb_tx_vlan_entry_test)
+define_ovsp4rt_test(fdb_rx_vlan_table_test)
+define_ovsp4rt_test(fdb_smac_table_test)
+define_ovsp4rt_test(fdb_tx_vlan_table_test)
 
-define_ipv4_tunnel_test(geneve_encap_v4_table_entry_test)
-define_ipv6_tunnel_test(geneve_encap_v6_table_entry_test)
-define_ipv4_tunnel_test(geneve_encap_v4_vlan_pop_test)
+define_ipv4_tunnel_test(geneve_encap_v4_table_test)
+define_ipv6_tunnel_test(geneve_encap_v6_table_test)
+define_ipv4_tunnel_test(geneve_encap_vlan_pop_v4_table_test)
 
-define_ovsp4rt_test(l2_to_v4_tunnel_test)
-define_ovsp4rt_test(l2_to_v6_tunnel_test)
+define_ovsp4rt_test(l2_to_tunnel_v4_table_test)
+define_ovsp4rt_test(l2_to_tunnel_v6_table_test)
 
 define_ovsp4rt_test(vxlan_decap_mod_entry_test)
 
-define_ipv4_tunnel_test(vxlan_encap_v4_table_entry_test)
-define_ipv6_tunnel_test(vxlan_encap_v6_table_entry_test)
-define_ipv4_tunnel_test(vxlan_encap_v4_vlan_pop_test)
-define_ipv6_tunnel_test(vxlan_encap_v6_vlan_pop_test)
+define_ipv4_tunnel_test(vxlan_encap_v4_table_test)
+define_ipv6_tunnel_test(vxlan_encap_v6_table_test)
+define_ipv4_tunnel_test(vxlan_encap_vlan_pop_v4_table_test)
+define_ipv6_tunnel_test(vxlan_encap_vlan_pop_v6_table_test)
 
 endif(ES2K_TARGET)
