+++ ovs-p4rt/sidecar/testing/fdb_rx_vlan_table_test.cc
@@ -1,14 +1,15 @@
 // Copyright 2024 Intel Corporation
 // SPDX-License-Identifier: Apache-2.0
 
-// Unit test for PrepareFdbSmacTableEntry()
+// Unit test for PrepareFdbRxVlanTableEntry()
 
-#include <arpa/inet.h>
 #include <stdint.h>
 
 #include <iostream>
 #include <string>
 
+#include "absl/flags/flag.h"
+#include "google/protobuf/util/json_util.h"
 #include "gtest/gtest.h"
 #include "logging/ovsp4rt_diag_detail.h"
 #include "ovsp4rt/ovs-p4rt.h"
@@ -18,8 +19,12 @@
 #include "p4info_text.h"
 #include "stratum/lib/utils.h"
 
+ABSL_FLAG(bool, dump_json, false, "Dump output table_entry in JSON");
+
 namespace ovsp4rt {
 
+using google::protobuf::util::JsonPrintOptions;
+using google::protobuf::util::MessageToJsonString;
 using stratum::ParseProtoFromString;
 
 constexpr bool INSERT_ENTRY = true;
@@ -27,9 +32,9 @@ constexpr bool REMOVE_ENTRY = false;
 
 static ::p4::config::v1::P4Info p4info;
 
-class FdbSmacEntryTest : public ::testing::Test {
+class FdbRxVlanTableTest : public ::testing::Test {
  protected:
-  FdbSmacEntryTest() {}
+  FdbRxVlanTableTest() { dump_json_ = absl::GetFlag(FLAGS_dump_json); }
 
   static void SetUpTestSuite() {
     ::util::Status status = ParseProtoFromString(P4INFO_TEXT, &p4info);
@@ -38,7 +43,7 @@ class FdbSmacEntryTest : public ::testing::Test {
     }
   }
 
-  void SetUp() { SelectTable("l2_fwd_smac_table"); }
+  void SetUp() { SelectTable("l2_fwd_rx_table"); }
 
   //----------------------------
   // P4Info lookup methods
@@ -65,6 +70,28 @@ class FdbSmacEntryTest : public ::testing::Test {
     return -1;
   }
 
+  int GetParamId(const std::string& param_name) const {
+    for (const auto& param : ACTION->params()) {
+      if (param.name() == param_name) {
+        return param.id();
+      }
+    }
+    std::cerr << "Action Param '" << param_name << "' not found!\n";
+    return -1;
+  }
+
+  void SelectAction(const std::string& action_name) {
+    for (const auto& action : p4info.actions()) {
+      const auto& pre = action.preamble();
+      if (pre.name() == action_name || pre.alias() == action_name) {
+        ACTION = &action;
+        ACTION_ID = pre.id();
+        return;
+      }
+    }
+    std::cerr << "Action '" << action_name << "' not found!\n";
+  }
+
   void SelectTable(const std::string& table_name) {
     for (const auto& table : p4info.tables()) {
       const auto& pre = table.preamble();
@@ -89,16 +116,31 @@ class FdbSmacEntryTest : public ::testing::Test {
     return word_value;
   }
 
+  void DumpTableEntry() {
+    if (dump_json_) {
+      JsonPrintOptions options;
+      options.add_whitespace = true;
+      options.preserve_proto_field_names = true;
+      std::string output;
+      ASSERT_TRUE(MessageToJsonString(table_entry, &output, options).ok());
+      std::cout << output << std::endl;
+    }
+  }
+
   //----------------------------
   // Initialization methods
   //----------------------------
 
+  void InitAction() { SelectAction("l2_fwd"); }
+
   void InitFdbInfo() {
     constexpr uint8_t MAC_ADDR[] = {0x11, 0x22, 0x33, 0x44, 0x55, 0x66};
-    constexpr uint8_t BRIDGE_ID = 42;
+    constexpr uint8_t BRIDGE_ID = 99;
+    constexpr uint32_t SRC_PORT = 0x42;
 
     memcpy(fdb_info.mac_addr, MAC_ADDR, sizeof(fdb_info.mac_addr));
     fdb_info.bridge_id = BRIDGE_ID;
+    fdb_info.rx_src_port = SRC_PORT;
   }
 
   //----------------------------
@@ -106,10 +148,23 @@ class FdbSmacEntryTest : public ::testing::Test {
   //----------------------------
 
   void CheckAction() const {
+    const int PORT_PARAM = GetParamId("port");
+
+    ASSERT_NE(ACTION_ID, -1);
+
     ASSERT_TRUE(table_entry.has_action());
     const auto& table_action = table_entry.action();
+
     const auto& action = table_action.action();
-    EXPECT_EQ(action.action_id(), GetActionId("NoAction"));
+    EXPECT_EQ(action.action_id(), ACTION_ID);
+
+    auto params = action.params();
+    ASSERT_EQ(action.params_size(), 1);
+
+    auto param = params[0];
+    ASSERT_EQ(param.param_id(), PORT_PARAM);
+    uint32_t port = DecodeWordValue(param.value());
+    EXPECT_EQ(port, fdb_info.rx_src_port);
   }
 
   void CheckBridgeId(const ::p4::v1::FieldMatch& match) const {
@@ -123,9 +178,7 @@ class FdbSmacEntryTest : public ::testing::Test {
     ASSERT_EQ(uint32_t(match_value[0]), uint32_t(fdb_info.bridge_id));
   }
 
-  void CheckDetail() const {
-    EXPECT_EQ(detail.table_id, LOG_L2_FWD_SMAC_TABLE);
-  }
+  void CheckDetail() const { EXPECT_EQ(detail.table_id, LOG_L2_FWD_RX_TABLE); }
 
   void CheckMacAddr(const ::p4::v1::FieldMatch& match) const {
     constexpr int MAC_ADDR_SIZE = 6;
@@ -135,21 +188,24 @@ class FdbSmacEntryTest : public ::testing::Test {
     ASSERT_EQ(match_value.size(), MAC_ADDR_SIZE);
 
     for (int i = 0; i < MAC_ADDR_SIZE; i++) {
-      EXPECT_EQ(match_value[i] & 0xFF, fdb_info.mac_addr[i])
+      EXPECT_EQ(match_value[i], fdb_info.mac_addr[i])
           << "mac_addr[" << i << "] is incorrect";
     }
   }
 
   void CheckMatches() const {
-    constexpr int MFID_MAC_ADDR = 1;
-    constexpr int MFID_BRIDGE_ID = 2;
+    constexpr char BRIDGE_ID_KEY[] = "user_meta.pmeta.bridge_id";
+    constexpr char DST_MAC_KEY[] = "dst_mac";
+
+    const int MFID_BRIDGE_ID = GetMatchFieldId(BRIDGE_ID_KEY);
+    const int MFID_DST_MAC = GetMatchFieldId(DST_MAC_KEY);
 
     // number of match fields
     ASSERT_EQ(table_entry.match_size(), 2);
 
     for (const auto& match : table_entry.match()) {
       auto field_id = match.field_id();
-      if (field_id == MFID_MAC_ADDR) {
+      if (field_id == MFID_DST_MAC) {
         CheckMacAddr(match);
       } else if (field_id == MFID_BRIDGE_ID) {
         CheckBridgeId(match);
@@ -159,16 +215,8 @@ class FdbSmacEntryTest : public ::testing::Test {
     }
   }
 
-  void CheckNoAction() const {
-    // Table entry does not specify an action
-    EXPECT_FALSE(table_entry.has_action());
-  }
-
   void CheckTableEntry() const {
-    // Table is defined (sanity check)
     ASSERT_FALSE(TABLE == nullptr);
-
-    // Table ID is what we expect
     EXPECT_EQ(table_entry.table_id(), TABLE_ID);
   }
 
@@ -183,6 +231,7 @@ class FdbSmacEntryTest : public ::testing::Test {
 
   // Comparison variables
   uint32_t TABLE_ID;
+  int ACTION_ID = -1;
 
  private:
   //----------------------------
@@ -190,34 +239,39 @@ class FdbSmacEntryTest : public ::testing::Test {
   //----------------------------
 
   const ::p4::config::v1::Table* TABLE = nullptr;
+  const ::p4::config::v1::Action* ACTION = nullptr;
+
+  bool dump_json_ = false;
 };
 
 //----------------------------------------------------------------------
-// PrepareFdbSmacTableEntry
+// PrepareFdbRxVlanTableEntry()
 //----------------------------------------------------------------------
 
-TEST_F(FdbSmacEntryTest, remove_entry) {
+TEST_F(FdbRxVlanTableTest, remove_entry) {
   // Arrange
   InitFdbInfo();
 
   // Act
-  PrepareFdbSmacTableEntry(&table_entry, fdb_info, p4info, REMOVE_ENTRY,
-                           detail);
+  PrepareFdbRxVlanTableEntry(&table_entry, fdb_info, p4info, REMOVE_ENTRY,
+                             detail);
+  DumpTableEntry();
 
   // Assert
   CheckDetail();
   CheckTableEntry();
   CheckMatches();
-  CheckNoAction();
 }
 
-TEST_F(FdbSmacEntryTest, insert_entry) {
+TEST_F(FdbRxVlanTableTest, insert_entry) {
   // Arrange
   InitFdbInfo();
+  InitAction();
 
   // Act
-  PrepareFdbSmacTableEntry(&table_entry, fdb_info, p4info, INSERT_ENTRY,
-                           detail);
+  PrepareFdbRxVlanTableEntry(&table_entry, fdb_info, p4info, INSERT_ENTRY,
+                             detail);
+  DumpTableEntry();
 
   // Assert
   CheckAction();
