+++ ovs-p4rt/sidecar/testing/vxlan_encap_v4_table_test.cc
@@ -7,43 +7,44 @@
 #include "gtest/gtest.h"
 #include "ovsp4rt/ovs-p4rt.h"
 #include "ovsp4rt_private.h"
-#include "testing/ipv6_tunnel_test.h"
+#include "testing/ipv4_tunnel_test.h"
 
 namespace ovsp4rt {
 
+class VxlanEncapV4TableTest : public Ipv4TunnelTest {};
+
 //----------------------------------------------------------------------
-// PrepareV6VxlanEncapAndVlanPopTableEntry()
+// PrepareVxlanEncapTableEntry()
 //----------------------------------------------------------------------
 
-TEST_F(Ipv6TunnelTest, vxlan_encap_v6_vlan_pop_params_are_correct) {
+TEST_F(VxlanEncapV4TableTest, minimal) {
   struct tunnel_info tunnel_info = {0};
   p4::v1::TableEntry table_entry;
   constexpr bool insert_entry = true;
 
-  constexpr uint32_t TABLE_ID = 34318005U;
-  constexpr uint32_t ACTION_ID = 28284062U;
+  constexpr uint32_t TABLE_ID = 40763773U;
+  constexpr uint32_t ACTION_ID = 20733968U;
 
   enum {
-    SRC_PORT_PARAM_ID = 7,
-    DST_PORT_PARAM_ID = 8,
-    VNI_PARAM_ID = 9,
+    SRC_PORT_PARAM_ID = 3,
+    DST_PORT_PARAM_ID = 4,
+    VNI_PARAM_ID = 5,
   };
 
   // Arrange
-  InitV6TunnelInfo(tunnel_info, OVS_TUNNEL_VXLAN);
+  InitV4TunnelInfo(tunnel_info, OVS_TUNNEL_VXLAN);
 
   // Act
-  PrepareV6VxlanEncapAndVlanPopTableEntry(&table_entry, tunnel_info, p4info,
-                                          insert_entry);
+  PrepareVxlanEncapTableEntry(&table_entry, tunnel_info, p4info, insert_entry);
   DumpTableEntry(table_entry);
 
   // Assert
-  EXPECT_EQ(table_entry.table_id(), TABLE_ID);
-
+  ASSERT_EQ(table_entry.table_id(), TABLE_ID);
   ASSERT_TRUE(table_entry.has_action());
+
   auto table_action = table_entry.action();
   auto action = table_action.action();
-  EXPECT_EQ(action.action_id(), ACTION_ID);
+  ASSERT_EQ(action.action_id(), ACTION_ID);
 
   auto params = action.params();
   int num_params = action.params_size();
@@ -56,7 +57,6 @@ TEST_F(Ipv6TunnelTest, vxlan_encap_v6_vlan_pop_params_are_correct) {
     auto param = params[i];
     int param_id = param.param_id();
     auto param_value = param.value();
-
     if (param_id == SRC_PORT_PARAM_ID) {
       src_port = DecodePortValue(param_value);
     } else if (param_id == DST_PORT_PARAM_ID) {
@@ -66,6 +66,7 @@ TEST_F(Ipv6TunnelTest, vxlan_encap_v6_vlan_pop_params_are_correct) {
     }
   }
 
+#if defined(ES2K_TARGET)
   ASSERT_TRUE(src_port.has_value());
 
   if (check_src_port_) {
@@ -74,6 +75,7 @@ TEST_F(Ipv6TunnelTest, vxlan_encap_v6_vlan_pop_params_are_correct) {
     // set the src_port param to (dst_port * 2).
     EXPECT_EQ(src_port.value(), DST_PORT * 2);  // SRC_PORT
   }
+#endif
 
   ASSERT_TRUE(dst_port.has_value());
   EXPECT_EQ(dst_port.value(), DST_PORT);
